requirements.txt
allure-pytest
faker
pytest
requests